Capstone_Crypto_Research/app/core/config.py
import os
from dotenv import load_dotenv
from pathlib import Path

env_path = Path(__file__).resolve().parent.parent / ".env"
load_dotenv(env_path)

NEWS_API_KEY = os.getenv("NEWS_API_KEY")
NEWS_API_URL = "https://newsapi.org/v2/everything"

if not NEWS_API_KEY:
    raise RuntimeError("NEWS_API_KEY not set")
