Capstone_Crypto_Research/requirements.txt
fastapi
uvicorn
qdrant-client
pytest
httpx
python-dotenv
requests